pyproject.toml
[project]
name = "agentic-rag-mcp-2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "faiss-cpu>=1.11.0",
    "fastmcp>=2.9.2",
    "langchain-cohere>=0.4.4",
    "langchain-google-genai>=2.1.6",
    "langchain-groq>=0.3.4",
    "langchain-mcp-adapters>=0.1.8",
    "langchain-openai>=0.3.27",
    "langgraph>=0.5.0",
    "langsmith>=0.4.4",
    "mysql-connector-python>=9.3.0",
    "openai>=1.93.0",
    "pandas>=2.3.0",
    "pip>=25.1.1",
    "pyfiglet>=1.0.3",
    "pymysql>=1.1.1",
    "python-dotenv>=1.1.1",
    "pyyaml>=6.0.2",
    "rich>=14.0.0",
    "sqlalchemy>=2.0.41",
    "yaspin>=3.1.0",
]
